RangeSlider optimization
