+++ best_flutter_ui_templates/lib/hotelBooking/SliderView.dart
@@ -1,0 +1,125 @@
+import 'package:flutter/material.dart';
+
+import 'hotelAppTheme.dart';
+
+class SliderView extends StatefulWidget {
+  final Function(double) onChnagedistValue;
+  final double distValue;
+
+  const SliderView({Key key, this.onChnagedistValue, this.distValue}) : super(key: key);
+  @override
+  _SliderViewState createState() => _SliderViewState();
+}
+
+class _SliderViewState extends State<SliderView> {
+  double distValue = 50.0;
+
+  @override
+  void initState() {
+    distValue = widget.distValue;
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Column(
+        children: <Widget>[
+          Row(
+            children: <Widget>[
+              Expanded(
+                flex: distValue.round(),
+                child: SizedBox(),
+              ),
+              Container(
+                width: 170,
+                child: Text(
+                  "Less than ${(distValue / 10).toStringAsFixed(1)} Km",
+                  textAlign: TextAlign.center,
+                ),
+              ),
+              Expanded(
+                flex: 100 - distValue.round(),
+                child: SizedBox(),
+              ),
+            ],
+          ),
+          SliderTheme(
+            data: SliderThemeData(
+              thumbShape: CustomThumbShape(),
+            ),
+            child: Slider(
+              onChanged: (value) {
+                setState(() {
+                  distValue = value;
+                });
+                try {
+                  widget.onChnagedistValue(distValue);
+                } catch (e) {}
+              },
+              min: 0,
+              max: 100,
+              activeColor: HotelAppTheme.buildLightTheme().primaryColor,
+              inactiveColor: Colors.grey.withOpacity(0.4),
+              divisions: 100,
+              value: distValue,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class CustomThumbShape extends SliderComponentShape {
+  static const double _thumbSize = 3.0;
+  static const double _disabledThumbSize = 3.0;
+
+  @override
+  Size getPreferredSize(bool isEnabled, bool isDiscrete) {
+    return isEnabled ? const Size.fromRadius(_thumbSize) : const Size.fromRadius(_disabledThumbSize);
+  }
+
+  static final Animatable<double> sizeTween = Tween<double>(
+    begin: _disabledThumbSize,
+    end: _thumbSize,
+  );
+
+  @override
+  void paint(
+    PaintingContext context,
+    Offset thumbCenter, {
+    Animation<double> activationAnimation,
+    Animation<double> enableAnimation,
+    bool isDiscrete,
+    TextPainter labelPainter,
+    RenderBox parentBox,
+    SliderThemeData sliderTheme,
+    TextDirection textDirection,
+    double value,
+  }) {
+    final Canvas canvas = context.canvas;
+    final ColorTween colorTween = ColorTween(
+      begin: sliderTheme.disabledThumbColor,
+      end: sliderTheme.thumbColor,
+    );
+    canvas.drawPath(
+        Path()
+          ..addOval(Rect.fromPoints(Offset(thumbCenter.dx + 12, thumbCenter.dy + 12), Offset(thumbCenter.dx - 12, thumbCenter.dy - 12)))
+          ..fillType = PathFillType.evenOdd,
+        Paint()
+          ..color = Colors.black.withOpacity(0.5)
+          ..maskFilter = MaskFilter.blur(BlurStyle.normal, convertRadiusToSigma(8)));
+
+    var cPaint = new Paint();
+    cPaint..color = Colors.white;
+    cPaint..strokeWidth = 14 / 2;
+    canvas.drawCircle(Offset(thumbCenter.dx, thumbCenter.dy), 12, cPaint);
+    cPaint..color = colorTween.evaluate(enableAnimation);
+    canvas.drawCircle(Offset(thumbCenter.dx, thumbCenter.dy), 10, cPaint);
+  }
+
+  double convertRadiusToSigma(double radius) {
+    return radius * 0.57735 + 0.5;
+  }
+}
